move and refactor user callback functions
