+++ packages/lesswrong/server.ts
@@ -89,7 +89,6 @@ import './server/callbacks/tagRelsCallbacks';
 import './server/staticRoutes/debugHeaders';
 import './server/tableOfContents';
 import './server/callbacks/subscriptionCallbacks';
-import './server/callbacks/rateLimitCallbacks';
 import './server/callbacks/reviewVoteCallbacks';
 import './server/callbacks/tagFlagCallbacks';
 import './server/callbacks/moderatorActionCallbacks';
@@ -99,8 +98,6 @@ import './server/callbacks/jargonTermCallbacks';
 
 import './server/resolvers/alignmentForumMutations';
 import './server/callbacks/alignment-forum/callbacks';
-import './server/callbacks/alignment-forum/alignmentPostCallbacks';
-import './server/callbacks/alignment-forum/alignmentUserCallbacks';
 import './server/callbacks/votingCallbacks';
 import './server/callbacks/electionCandidateCallbacks';
