+++ packages/lesswrong/server/search/elastic/elasticCallbacks.ts
@@ -3,7 +3,6 @@ import {
   searchIndexedCollectionNames,
 } from "../../../lib/search/searchUtil";
 import { getCollectionHooks } from "../../mutationCallbacks";
-import UsersRepo from "../../repos/UsersRepo";
 import ElasticClient from "./ElasticClient";
 import ElasticExporter from "./ElasticExporter";
 import { isElasticEnabled } from "../../../lib/instanceSettings";
@@ -32,29 +31,4 @@ export const registerElasticCallbacks = () => {
     getCollectionHooks(collectionName).createAsync.add(({ document }) => callback(document));
     getCollectionHooks(collectionName).editAsync.add(callback);
   }
-
-  // editAsync
-  async function reindexDeletedUserContent(newUser: DbUser, oldUser: DbUser) {
-    if (!!newUser.deleted !== !!oldUser.deleted) {
-      const repo = new UsersRepo();
-      const [
-        postIds,
-        commentIds,
-        sequenceIds,
-      ] = await Promise.all([
-        repo.getAllUserPostIds(newUser._id),
-        repo.getAllUserCommentIds(newUser._id),
-        repo.getAllUserSequenceIds(newUser._id),
-      ]);
-
-      const client = new ElasticClient();
-      const exporter = new ElasticExporter(client);
-      await Promise.all([
-        ...postIds.map((id) => exporter.updateDocument("Posts", id)),
-        ...commentIds.map((id) => exporter.updateDocument("Comments", id)),
-        ...sequenceIds.map((id) => exporter.updateDocument("Sequences", id)),
-      ]);
-    }
-  }
-  getCollectionHooks("Users").editAsync.add(reindexDeletedUserContent);
 }
