+++ packages/lesswrong/server/voteServer.ts
@@ -5,7 +5,7 @@ import { recalculateScore } from '../lib/scoring';
 import { isValidVoteType } from '../lib/voting/voteTypes';
 import { VoteDocTuple, getVotePower } from '../lib/voting/vote';
 import { getVotingSystemForDocument, VotingSystem } from '../lib/voting/votingSystems';
-import { createAnonymousContext } from './vulcan-lib/query';
+import { createAdminContext, createAnonymousContext } from './vulcan-lib/query';
 import { randomId } from '../lib/random';
 import { getConfirmedCoauthorIds } from '../lib/collections/posts/helpers';
 import { ModeratorActions } from '../lib/collections/moderatorActions/collection';
@@ -25,7 +25,11 @@ import VotesRepo from './repos/VotesRepo';
 import { swrInvalidatePostRoute } from './cache/swr';
 import { onCastVoteAsync, onVoteCancel } from './callbacks/votingCallbacks';
 import { getVoteAFPower } from './callbacks/alignment-forum/callbacks';
-import { isElasticEnabled, isLWorAF } from "../lib/instanceSettings";
+import { isElasticEnabled } from "../lib/instanceSettings";
+import { getCollection } from '@/lib/vulcan-lib/getCollection';
+import Users from '@/lib/collections/users/collection';
+import { capitalize } from '@/lib/vulcan-lib/utils';
+import { getVoteableCollections } from '@/lib/make_voteable';
 
 // Test if a user has voted on the server
 const getExistingVote = async ({ document, user }: {
@@ -630,4 +634,44 @@ export const recalculateDocumentScores = async (document: VoteableType, collecti
     afExtendedScore: await votingSystem.computeExtendedScore(afVotes, context),
     score: recalculateScore({...document, baseScore})
   };
+}
+
+
+/**
+ * Reverse the given vote, without triggering any karma change notifications
+ */
+export async function silentlyReverseVote(vote: DbVote, context: ResolverContext) {
+  const collection = getCollection(vote.collectionName as VoteableCollectionName);
+  const document = await collection.findOne({_id: vote.documentId});
+  const user = await Users.findOne({_id: vote.userId});
+  if (document && user) {
+    await clearVotesServer({ document, collection, user, silenceNotification: true, context });
+  } else {
+    //eslint-disable-next-line no-console
+    console.info("No item or user found corresponding to vote: ", vote, document, user);
+  }
+}
+
+async function nullifyVotesForUserAndCollection(user: DbUser, collection: CollectionBase<VoteableCollectionName>) {
+  const collectionName = capitalize(collection.collectionName);
+  const context = createAdminContext();
+  const votes = await Votes.find({
+    collectionName: collectionName,
+    userId: user._id,
+    cancelled: false,
+  }).fetch();
+  for (let vote of votes) {
+    //eslint-disable-next-line no-console
+    console.log("reversing vote: ", vote)
+    await silentlyReverseVote(vote, context);
+  };
+  //eslint-disable-next-line no-console
+  console.info(`Nullified ${votes.length} votes for user ${user.username}`);
+}
+
+
+export async function nullifyVotesForUser(user: DbUser) {
+  for (let collection of getVoteableCollections()) {
+    await nullifyVotesForUserAndCollection(user, collection);
+  }
 }
